Create trainticketsearcher.py
This is a train ticket searcher in Python 3.4. 
It looks for trains heading from one start to one destination on a series of sequential dates,
with the information of the number of remaining tickets & the start/arrive time & the type of seat & all the information to know about the trip on train.
This is a final project for my python class at school.